requirements.txt
requests~=2.32.3
terminaltables~=3.1.10
python-dotenv~=1.0.1